requirements.txt
asgiref==3.11.1
crispy-bootstrap5==2026.3
Django==6.0.4
django-crispy-forms==2.6
django-filter==25.2
pillow==12.2.0
psycopg2-binary==2.9.12
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.2
